Project_I/requirements.txt
kagglehub==0.3.10
numpy==2.2.3
pillow==11.1.0
torch==2.6.0